Extract benchmark_common.hpp and add logfmt large benchmark
Move the shared streaming-benchmark scaffolding (timing, structural-byte
accounting, peak-working-set sampling, the AppendBatch sink, and the
release-build gate) out of benchmark_json.cpp into a header-only
`bench` namespace. `RunStreamingFlow` / `RunStreamingBenchmark` now take
a `ParserStreamFn` callable so any parser plugs into the same harness.
Consolidates benchmark_stream.cpp's duplicated release-build gate too.

Add benchmark_logfmt.cpp with the 1M-line `[logfmt_parser][large]`
streaming case, fed by the same RNG-driven records as JSON so lines/s is
directly comparable. Verified in a release build: logfmt streams 1M rows
correctly at throughput on par with JSON.
